requirements.txt
# Core Dashboard
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
pyarrow==15.0.0

# AI/ML
openai==1.12.0
anthropic==0.18.1
langchain==0.1.6
langchain-openai==0.0.5

# Data Processing
pyyaml==6.0.1
python-dateutil==2.8.2
faker==22.6.0

# Utilities
python-dotenv==1.0.1
